+++ app/Providers/AppServiceProvider.php
@@ -21,7 +21,7 @@ class AppServiceProvider extends ServiceProvider
     {
         \App\Models\Finance::observe(\App\Observers\AuditObserver::class);
         \App\Models\Document::observe(\App\Observers\AuditObserver::class);
-        \App\Models\Meeting::observe(\App\Observers\AuditObserver::class);
+        \App\Models\Agenda::observe(\App\Observers\AuditObserver::class);
         \App\Models\Warning::observe(\App\Observers\AuditObserver::class);
         \App\Models\Event::observe(\App\Observers\AuditObserver::class);
         \App\Models\User::observe(\App\Observers\AuditObserver::class);